+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.5.0] - 2018-07-15
+### Fixed
+Some styles in Trello add card popup
+
+### Added
+Profile contacts
+
 ## [1.4.0] - 2018-07-12
 ### Fixed
 Some little bugs